refactor: 패키지 구조 변경, 접근제어자 변경
