+++ src/main/java/blackjack/Blackjack.java
@@ -1,12 +1,12 @@
 package blackjack;
 
-import blackjack.domain.card.Deck;
+import blackjack.domain.playing.HitOrStay;
+import blackjack.domain.playing.deck.Deck;
+import blackjack.domain.playing.user.Dealer;
+import blackjack.domain.playing.user.Player;
+import blackjack.domain.playing.user.Players;
+import blackjack.domain.playing.user.exception.HitOrStayException;
 import blackjack.domain.result.GameResult;
-import blackjack.domain.user.Dealer;
-import blackjack.domain.user.HitOrStay;
-import blackjack.domain.user.Player;
-import blackjack.domain.user.Players;
-import blackjack.domain.user.exception.HitOrStayException;
 import blackjack.view.InputView;
 import blackjack.view.OutputView;
