--- src/main/java/blackjack/domain/card/exception/DeckException.java
@@ -1,7 +1,0 @@
-package blackjack.domain.card.exception;
-
-public class DeckException extends RuntimeException {
-    public DeckException(String message) {
-        super(message);
-    }
-}
