--- src/main/java/blackjack/domain/card/exception/ScoreException.java
@@ -1,7 +1,0 @@
-package blackjack.domain.card.exception;
-
-public class ScoreException extends RuntimeException {
-    public ScoreException(String message) {
-        super(message);
-    }
-}
