+++ src/main/java/blackjack/domain/playing/HitOrStay.java
@@ -1,0 +1,28 @@
+package blackjack.domain.playing;
+
+import blackjack.domain.playing.user.exception.HitOrStayException;
+
+import java.util.Arrays;
+
+public enum HitOrStay {
+
+    HIT("y"),
+    STAY("n");
+
+    private String input;
+
+    HitOrStay (String input) {
+        this.input = input;
+    }
+
+    public static HitOrStay of(String input) {
+        return Arrays.stream(values())
+                .filter(hitOrStay -> hitOrStay.input.equals(input))
+                .findFirst()
+                .orElseThrow(HitOrStayException::new);
+    }
+
+    public boolean isToHit() {
+        return this == HIT;
+    }
+}
