+++ src/main/java/blackjack/domain/playing/card/Card.java
@@ -1,0 +1,47 @@
+package blackjack.domain.playing.card;
+
+import java.util.Objects;
+
+public class Card {
+    private final Symbol symbol;
+    private final Type type;
+
+    public Card(Symbol symbol, Type type) {
+        this.symbol = symbol;
+        this.type = type;
+    }
+
+    public boolean isAce() {
+        return symbol == Symbol.ACE;
+    }
+
+    public Score getScore() {
+        return symbol.getScore();
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        Card card = (Card) o;
+        return symbol == card.symbol &&
+                type == card.type;
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(symbol, type);
+    }
+
+    @Override
+    public String toString() {
+        return "Card{" +
+                "symbol=" + symbol +
+                ", type=" + type +
+                '}';
+    }
+
+    public String getName() {
+        return symbol.getName() + " " + type.getName();
+    }
+}
