+++ src/main/java/blackjack/domain/playing/card/exception/ScoreException.java
@@ -1,0 +1,7 @@
+package blackjack.domain.playing.card.exception;
+
+public class ScoreException extends RuntimeException {
+    public ScoreException(String message) {
+        super(message);
+    }
+}
