+++ src/main/java/blackjack/domain/playing/deck/exception/DeckException.java
@@ -1,0 +1,7 @@
+package blackjack.domain.playing.deck.exception;
+
+public class DeckException extends RuntimeException {
+    public DeckException(String message) {
+        super(message);
+    }
+}
