+++ src/main/java/blackjack/domain/playing/user/Player.java
@@ -1,0 +1,11 @@
+package blackjack.domain.playing.user;
+
+public class Player extends User {
+    private Player(String name) {
+        super(name, Cards.emptyCards());
+    }
+
+    public static Player of(String name) {
+        return new Player(name);
+    }
+}
