+++ src/main/java/blackjack/domain/playing/user/exception/HitOrStayException.java
@@ -1,0 +1,7 @@
+package blackjack.domain.playing.user.exception;
+
+public class HitOrStayException extends RuntimeException {
+    public HitOrStayException() {
+        super("유효하지않는 입력입니다.");
+    }
+}
