+++ src/main/java/blackjack/domain/playing/user/exception/PlayersException.java
@@ -1,0 +1,7 @@
+package blackjack.domain.playing.user.exception;
+
+public class PlayersException extends RuntimeException {
+	public PlayersException(String message) {
+		super(message);
+	}
+}
