+++ src/main/java/blackjack/domain/playing/user/exception/UserException.java
@@ -1,0 +1,7 @@
+package blackjack.domain.playing.user.exception;
+
+public class UserException extends RuntimeException {
+    public UserException(String message) {
+        super(message);
+    }
+}
