+++ src/main/java/blackjack/domain/result/GameResult.java
@@ -1,8 +1,8 @@
 package blackjack.domain.result;
 
-import blackjack.domain.user.Dealer;
-import blackjack.domain.user.Player;
-import blackjack.domain.user.Players;
+import blackjack.domain.playing.user.Dealer;
+import blackjack.domain.playing.user.Player;
+import blackjack.domain.playing.user.Players;
 
 import java.util.Arrays;
 import java.util.Collections;
