--- src/main/java/blackjack/domain/user/Player.java
@@ -1,11 +1,0 @@
-package blackjack.domain.user;
-
-public class Player extends User {
-    private Player(String name) {
-        super(name, Cards.emptyCards());
-    }
-
-    public static Player of(String name) {
-        return new Player(name);
-    }
-}
