--- src/main/java/blackjack/domain/user/exception/HitOrStayException.java
@@ -1,7 +1,0 @@
-package blackjack.domain.user.exception;
-
-public class HitOrStayException extends RuntimeException {
-    public HitOrStayException() {
-        super("유효하지않는 입력입니다.");
-    }
-}
