--- src/main/java/blackjack/domain/user/exception/PlayersException.java
@@ -1,7 +1,0 @@
-package blackjack.domain.user.exception;
-
-public class PlayersException extends RuntimeException {
-	public PlayersException(String message) {
-		super(message);
-	}
-}
