--- src/main/java/blackjack/domain/user/exception/UserException.java
@@ -1,7 +1,0 @@
-package blackjack.domain.user.exception;
-
-public class UserException extends RuntimeException {
-    public UserException(String message) {
-        super(message);
-    }
-}
