+++ src/test/java/blackjack/domain/result/GameResultTest.java
@@ -1,12 +1,12 @@
 package blackjack.domain.result;
 
-import blackjack.domain.card.Card;
-import blackjack.domain.card.Deck;
-import blackjack.domain.card.Symbol;
-import blackjack.domain.card.Type;
-import blackjack.domain.user.Dealer;
-import blackjack.domain.user.Player;
-import blackjack.domain.user.Players;
+import blackjack.domain.playing.card.Card;
+import blackjack.domain.playing.card.Symbol;
+import blackjack.domain.playing.card.Type;
+import blackjack.domain.playing.deck.Deck;
+import blackjack.domain.playing.user.Dealer;
+import blackjack.domain.playing.user.Player;
+import blackjack.domain.playing.user.Players;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mock;
